RAG_Trials/test9/requirements.txt
# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-oauth2==1.1.1

# Database  
duckdb==0.9.2

# Vector operations
numpy==1.24.3
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# HTTP client
httpx==0.25.2

# Monitoring
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0